Bring Graph500 code up to current version.
Introduces lightweight BFS tasks (taking only int64's), separate
options file, and some stuff for reading and writing checkpoints
to save on computation for everything except bfs itself.

+++ applications/graph500/softxmt/rmat.h
@@ -9,6 +9,6 @@ extern double C;
 extern double D;
 extern mrg_state * prng_state;
 
-void rmat_edgelist(tuple_graph* grin, int SCALE);
+void rmat_edgelist(tuple_graph* grin, int64_t SCALE);
 
 #endif /* SOFTXMT_RMAT_H */
